keelson-controller-gamepad/requirements.txt
pyusb==1.2.1
requests==2.32.3
keelson==0.4.4
eclipse-zenoh==1.6.2
